move config to zelback folder
