+++ ZelBack/server.js
@@ -1,4 +1,5 @@
 // ZelBack configuration
+process.env.NODE_CONFIG_DIR = `${__dirname}/config/`;
 const config = require('config');
 // const fs = require('fs');
 // const https = require('https');
