+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "zelflux",
-  "version": "0.44.0",
+  "version": "0.45.0",
   "description": "ZelFlux - ZelNode Daemon. The entrace to the Zel network.",
   "repository": {
     "type": "git",
@@ -68,6 +68,7 @@
     "path": "^0.12.7",
     "qs": "^6.9.1",
     "store": "^2.0.12",
+    "stream": "0.0.2",
     "vue": "^2.6.11",
     "vue-router": "^3.1.3",
     "vuex": "^3.1.2",
